+++ Editor/Component/Window_ComponentEditor.cs
@@ -1,37 +1,32 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using HananokiEditor.Extensions;
 using UnityEditor;
 using UnityEngine;
-using System.Collections.Generic;
-using HananokiEditor.Extensions;
-using HananokiRuntime.Extensions;
-using System;
-using System.Linq;
-using UnityEditor;
-using UnityEditor.SceneManagement;
-using UnityEditorInternal;
-using UnityEngine;
-using UnityReflection;
 using UnityObject = UnityEngine.Object;
 
 namespace HananokiEditor.CustomHierarchy {
-	public class ComponentEditor : HNEditorWindow<ComponentEditor> {
+	public class Window_ComponentEditor : HNEditorWindow<Window_ComponentEditor> {
 		UnityObject m_component;
 		bool inited;
 		Vector2 m_scroll;
 		public Editor m_currentEditor;
 
+
+		/////////////////////////////////////////
 		public static void Open( UnityObject component ) {
-			var window = GetWindow<ComponentEditor>();
+			var window = EditorWindowUtils.ShowWindow<Window_ComponentEditor>();
 			window.m_component = component;
 		}
 
+
+		/////////////////////////////////////////
 		void Init() {
 			SetTitle( m_component.GetType().Name, m_component.GetCachedIcon() );
-			m_currentEditor = Editor.CreateEditor( m_component );
+			m_currentEditor = ComponentEditorCache.Get( m_component );
+			inited = true;
 		}
 
+
+		/////////////////////////////////////////
 		public override void OnDefaultGUI() {
 			if( !inited ) Init();