+++ Editor/UI/TreeView_Components.cs
@@ -1,17 +1,12 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-using HananokiEditor.Extensions;
-using HananokiRuntime;
-using HananokiRuntime.Extensions;
+﻿using HananokiEditor.Extensions;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using UnityEditor;
 using UnityEditor.IMGUI.Controls;
-using UnityEngine.UI;
-
+using UnityEngine;
 using E = HananokiEditor.CustomHierarchy.SettingsEditor;
-using UnityObject = UnityEngine.Object;
+
 
 namespace HananokiEditor.CustomHierarchy {
 	using Item = TreeView_Components.Item;
@@ -19,7 +14,6 @@ namespace HananokiEditor.CustomHierarchy {
 	public sealed class TreeView_Components : HTreeView<Item> {
 
 		const int kMove = 0;
-		//const int kEnable = 1;
 		const int kComponents = 1;
 		const int kSearch = 2;
 		const int kInspector = 3;
@@ -30,6 +24,8 @@ namespace HananokiEditor.CustomHierarchy {
 		}
 
 
+
+		/////////////////////////////////////////
 		public TreeView_Components() : base( new TreeViewState() ) {
 			E.Load();
 
@@ -75,9 +71,12 @@ namespace HananokiEditor.CustomHierarchy {
 		}
 
 
+
+		/////////////////////////////////////////
 		public void RegisterFiles() {
 
-			var lst = new List<Item>();
+			m_root = new Item { depth = -1 };
+
 			InitID();
 
 			foreach( var p in E.i.m_componentHandlerData ) {
@@ -89,31 +88,26 @@ namespace HananokiEditor.CustomHierarchy {
 					icon = p.type.GetIcon(),
 					data = p,
 				};
-				lst.Add( item );
+				m_root.AddChild( item );
 			}
 
-			m_registerItems = lst;
 			ReloadAndSorting();
 		}
 
 
+
+		/////////////////////////////////////////
 		public void ReloadAndSorting() {
 			Reload();
+			//RollbackLastSelect();
 		}
 
 
 
+		/////////////////////////////////////////
 		protected override void OnRowGUI( RowGUIArgs args ) {
 			var item = (Item) args.item;
 			bool changed = false;
-
-
-
-			//if( item.currentEnabled && !args.selected ) {
-			//	var col1 = ColorUtils.RGB( 169, 201, 255 );
-			//	col1.a = 0.5f;
-			//	EditorGUI.DrawRect( args.rowRect, col1 );
-			//}
 
 			for( var i = 0; i < args.GetNumVisibleColumns(); i++ ) {
 				var rect = args.GetCellRect( i );
@@ -126,15 +120,6 @@ namespace HananokiEditor.CustomHierarchy {
 					EditorGUI.LabelField( rect.AlignCenterH( 16 ), EditorHelper.TempContent( EditorIcon.toolbar_minus ) );
 					break;
 				}
-				//case kEnable: {
-				//	ScopeChange.Begin();
-				//	var _b = EditorGUI.Toggle( rect.AlignCenter( 16, 16 ), item.data.enable );
-				//	if( ScopeChange.End() ) {
-				//		item.data.enable = _b;
-				//		E.Save();
-				//	}
-				//	break;
-				//}
 				case kComponents: {
 					Label( args, rect, item.displayName, item.icon );
 					break;
@@ -168,5 +153,97 @@ namespace HananokiEditor.CustomHierarchy {
 				EditorWindowUtils.RepaintHierarchyWindow();
 			}
 		}
+
+
+
+		/////////////////////////////////////////
+		protected override void SingleSelectionChanged( Item item ) {
+			BackupLastSelect( item );
+		}
+
+
+
+		#region DragAndDrop
+
+		class DragAndDropData {
+			public List<Item> dragItems;
+
+			public static DragAndDropVisualMode visualMode = DragAndDropVisualMode.None;
+			public DragAndDropArgs args;
+
+			//public Item dropItem => (Item) args.parentItem;
+
+			public DragAndDropData( string dragID, DragAndDropArgs args ) {
+				this.args = args;
+				dragItems = DragAndDrop.GetGenericData( dragID ) as List<Item>;
+			}
+
+			public List<Item> HandleBetweenItems( List<Item> items ) {
+				if( dragItems == null ) {
+					visualMode = DragAndDropVisualMode.None;
+					return null;
+				}
+
+				visualMode = DragAndDropVisualMode.Move;
+
+				// ドロップを実行します
+				if( !args.performDrop ) return null;
+
+				var lst = items.ToList();
+				var insertIndex = args.insertAtIndex;
+
+				foreach( var p in dragItems ) {
+					var idx = lst.FindIndex( x => x.id == p.id );
+					if( idx < insertIndex ) {
+						insertIndex--;
+					}
+					lst.Remove( p );
+				}
+
+				lst.InsertRange( insertIndex, dragItems );
+
+				return lst;
+			}
+		}
+
+
+		protected override DragAndDropVisualMode HandleDragAndDrop( DragAndDropArgs args ) {
+			var data = new DragAndDropData( dragID, args );
+
+			switch( args.dragAndDropPosition ) {
+			case DragAndDropPosition.BetweenItems:
+				var lst = data.HandleBetweenItems( m_root.children.OfType<Item>().ToList() );
+				if( lst != null ) {
+					E.i.m_componentHandlerData = lst.Select( x => x.data ).ToList();
+					E.Save();
+					RegisterFiles();
+					SelectItem( m_root.children.Find( x => x.displayName == data.dragItems[ 0 ].displayName ) );
+				}
+				break;
+			case DragAndDropPosition.UponItem:
+			case DragAndDropPosition.OutsideItems:
+				DragAndDropData.visualMode = DragAndDropVisualMode.None;
+				break;
+			}
+			return DragAndDropData.visualMode;
+		}
+
+
+		protected override void SetupDragAndDrop( SetupDragAndDropArgs args ) {
+			if( args.draggedItemIDs == null ) return;
+
+			DragAndDrop.PrepareStartDrag();
+
+			DragAndDrop.paths = null;
+			DragAndDrop.SetGenericData( dragID, ToItems( args.draggedItemIDs ).ToList() );
+			DragAndDrop.visualMode = DragAndDropVisualMode.None;
+			DragAndDrop.StartDrag( $"{dragID}.StartDrag" );
+		}
+
+		protected override bool CanStartDrag( CanStartDragArgs args ) {
+			return true;
+		}
+
+		#endregion
 	}
 }